requirements.txt
pyyaml>=6.0
pyarrow>=14.0.0
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23.0
mypy>=1.0
ib_async>=1.0.0
aiohttp>=3.9.0
